requirements.txt
# GitHub API
PyGithub==2.1.1

# HTTP requests
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Testing (Phase 2)
pytest==7.4.3
pytest-cov==4.1.0

# Data validation
pyyaml==6.0.1
